chore: move action directive to ha folder
